requirements-rl.txt
torch>=2.2.0
transformers>=4.44.0
peft>=0.12.0
trl>=0.10.1
accelerate>=0.34.0
sentencepiece>=0.2.0
faiss-cpu>=1.8.0
numpy>=1.26.0
tensorboard>=2.17.0
pytest>=8.3.0
